Move `CustomOAuth` module
