+++ apps/meteor/app/apple/client/index.ts
@@ -1,4 +1,4 @@
-import { CustomOAuth } from '../../custom-oauth/client/CustomOAuth';
+import { CustomOAuth } from '../../../client/lib/customOAuth/CustomOAuth';
 import { config } from '../lib/config';
 
 CustomOAuth.configureOAuthService('apple', config);
