+++ apps/meteor/app/github-enterprise/client/hooks/useGitHubEnterpriseAuth.ts
@@ -2,7 +2,7 @@ import type { OauthConfig } from '@rocket.chat/core-typings';
 import { useSetting } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
-import { CustomOAuth } from '../../../custom-oauth/client/CustomOAuth';
+import { CustomOAuth } from '../../../../client/lib/customOAuth/CustomOAuth';
 
 // GitHub Enterprise Server CallBack URL needs to be http(s)://{rocketchat.server}[:port]/_oauth/github_enterprise
 // In RocketChat -> Administration the URL needs to be http(s)://{github.enterprise.server}/
