+++ apps/meteor/app/tokenpass/client/hooks/useTokenPassAuth.tsx
@@ -2,7 +2,7 @@ import type { OauthConfig } from '@rocket.chat/core-typings';
 import { useSetting } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
-import { CustomOAuth } from '../../../custom-oauth/client/CustomOAuth';
+import { CustomOAuth } from '../../../../client/lib/customOAuth/CustomOAuth';
 
 const config = {
 	serverURL: '',
