+++ apps/meteor/client/lib/customOAuth/CustomOAuth.ts
@@ -5,11 +5,11 @@ import { Accounts } from 'meteor/accounts-base';
 import { Meteor } from 'meteor/meteor';
 import { OAuth } from 'meteor/oauth';
 
-import type { IOAuthProvider } from '../../../client/definitions/IOAuthProvider';
-import { overrideLoginMethod, type LoginCallback } from '../../../client/lib/2fa/overrideLoginMethod';
-import { loginServices } from '../../../client/lib/loginServices';
-import { createOAuthTotpLoginMethod } from '../../../client/meteorOverrides/login/oauth';
 import { isURL } from '../../../lib/utils/isURL';
+import type { IOAuthProvider } from '../../definitions/IOAuthProvider';
+import { createOAuthTotpLoginMethod } from '../../meteorOverrides/login/oauth';
+import { overrideLoginMethod, type LoginCallback } from '../2fa/overrideLoginMethod';
+import { loginServices } from '../loginServices';
 
 // Request custom OAuth credentials for the user
 // @param options {optional}
